refactor(remediation-mobile): split into sibling components for 500-line cap
`frontend/src/pages/Remediation/Mobile.tsx` was 700 lines, over the
500-line soft cap. Decomposed into:

- mobileTypes.ts        — types, constants, provider helpers
- InFlightTile.tsx      — in-flight dispatch tile (with elapsed timer)
- ActionSheet.tsx       — dispatch / agent-picker bottom sheet
- RemediationLists.tsx  — AutomationsList, PRsList, IssuesList

Main `Mobile.tsx` (now 262 lines) keeps data fetching, dispatch
plumbing, and composes the sub-components. Public exports
`RemediationMobile` and `InFlightDispatch` are preserved (the type is
re-exported from `mobileTypes`).

No behavior change.
